refactor: move unrelated tests from authentication repository to account repository test
